refactor to compose-only navigation system
