+++ app/src/main/java/com/codingwithmitch/food2forkcompose/presentation/ui/recipe_list/RecipeListScreen.kt
@@ -1,0 +1,72 @@
+package com.codingwithmitch.food2forkcompose.presentation.ui.recipe_list
+
+import android.util.Log
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.Scaffold
+import androidx.compose.material.rememberScaffoldState
+import androidx.compose.runtime.Composable
+import com.codingwithmitch.food2forkcompose.presentation.components.RecipeList
+import com.codingwithmitch.food2forkcompose.presentation.components.SearchAppBar
+import com.codingwithmitch.food2forkcompose.presentation.theme.AppTheme
+import com.codingwithmitch.food2forkcompose.util.TAG
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+
+@ExperimentalCoroutinesApi
+@ExperimentalMaterialApi
+@Composable
+fun RecipeListScreen(
+  isDarkTheme: Boolean,
+  onToggleTheme: () -> Unit,
+  viewModel: RecipeListViewModel,
+){
+  Log.d(TAG, "RecipeListScreen: ${viewModel}")
+  
+  val recipes = viewModel.recipes.value
+
+  val query = viewModel.query.value
+
+  val selectedCategory = viewModel.selectedCategory.value
+
+  val loading = viewModel.loading.value
+
+  val page = viewModel.page.value
+
+  val scaffoldState = rememberScaffoldState()
+
+  AppTheme(
+    displayProgressBar = loading,
+    scaffoldState = scaffoldState,
+    darkTheme = isDarkTheme,
+  ) {
+    Scaffold(
+      topBar = {
+        SearchAppBar(
+          query = query,
+          onQueryChanged = viewModel::onQueryChanged,
+          onExecuteSearch = {
+            viewModel.onTriggerEvent(RecipeListEvent.NewSearchEvent)
+          },
+          categories = getAllFoodCategories(),
+          selectedCategory = selectedCategory,
+          onSelectedCategoryChanged = viewModel::onSelectedCategoryChanged,
+          onToggleTheme = {onToggleTheme()}
+        )
+      },
+      scaffoldState = scaffoldState,
+      snackbarHost = {
+        scaffoldState.snackbarHostState
+      },
+    ) {
+      RecipeList(
+        loading = loading,
+        recipes = recipes,
+        onChangeScrollPosition = viewModel::onChangeRecipeScrollPosition,
+        page = page,
+        onTriggerNextPage = { viewModel.onTriggerEvent(RecipeListEvent.NextPageEvent) },
+        onNavigateToRecipeDetailScreen = {
+          TODO("navigate to RecipeScreen")
+        }
+      )
+    }
+  }
+}
