--- app/src/main/res/navigation/main_graph.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/main_graph"
-    app:startDestination="@id/recipeListFragment">
-
-    <fragment
-        android:id="@+id/recipeListFragment"
-        android:name="com.codingwithmitch.food2forkcompose.presentation.ui.recipe_list.RecipeListFragment"
-        android:label="RecipeListFragment">
-        <action
-            android:id="@+id/viewRecipe"
-            app:destination="@id/recipeFragment" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/recipeFragment"
-        android:name="com.codingwithmitch.food2forkcompose.presentation.ui.recipe.RecipeFragment"
-        android:label="RecipeFragment">
-    </fragment>
-</navigation>
